fix: explicitly use qemu:///system for virt-install
virt-install defaults to qemu:///session when run as non-root.
Must explicitly specify qemu:///system to use the system
libvirt where the default network is already active.

This is why we were getting 'network not active' errors
even though the system network WAS active - virt-install
was trying to use session libvirt.